Refactored the backend based on the new structure of main.
