+++ backend/apps/data_sources/admin.py
@@ -1,3 +1,22 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import DataSource
+
+
+@admin.register(DataSource)
+class DataSourceAdmin(admin.ModelAdmin):
+    list_display = (
+        "name",
+        "project",
+        "source_type",
+        "data_format",
+        "contains_personal_data",
+        "created_at",
+    )
+    list_filter = (
+        "source_type",
+        "data_format",
+        "contains_personal_data",
+        "project",
+    )
+    search_fields = ("name", "description", "location", "project__name")
