+++ backend/apps/projects/admin.py
@@ -1,3 +1,10 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import Project
+
+
+@admin.register(Project)
+class ProjectAdmin(admin.ModelAdmin):
+    list_display = ("name", "status", "created_at", "updated_at")
+    list_filter = ("status", "created_at")
+    search_fields = ("name", "description")
