+++ backend/config/urls.py
@@ -15,11 +15,12 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path
+from django.urls import include, path
+
 from apps.projects.views import health_check
 
 urlpatterns = [
     path("admin/", admin.site.urls),
+    path("api/", include("apps.data_sources.urls")),
     path("api/health/", health_check),
 ]
-
