--- backend/projects/admin.py
@@ -1,9 +1,0 @@
-from django.contrib import admin
-from .models import Project
-
-
-@admin.register(Project)
-class ProjectAdmin(admin.ModelAdmin):
-    list_display = ("name", "status", "created_at", "updated_at")
-    list_filter = ("status", "created_at")
-    search_fields = ("name", "description")
