--- backend/projects/models.py
@@ -1,23 +1,0 @@
-from django.db import models
-
-
-class Project(models.Model):
-    class Status(models.TextChoices):
-        ACTIVE = "active", "Active"
-        ARCHIVED = "archived", "Archived"
-
-    name = models.CharField(max_length=100)
-    description = models.TextField(blank=True)
-    status = models.CharField(
-        max_length=20,
-        choices=Status.choices,
-        default=Status.ACTIVE,
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        ordering = ["-created_at"]
-
-    def __str__(self):
-        return self.name
