--- backend/pyproject.toml
@@ -1,7 +1,0 @@
-[project]
-name = "backend"
-version = "0.1.0"
-requires-python = ">=3.14"
-dependencies = [
-    "django>=6.0.4",
-]
